+++ src/main/java/com/clothify/pos/dao/custom/impl/ProductDaoImpl.java
@@ -96,10 +96,10 @@ public class ProductDaoImpl implements ProductDao {
     }
 
     @Override
-    public int count(){
+    public long count(){
         Session session = HibernateUtil.getSession();
         session.getTransaction().begin();
-        int singleResult = (int) session.createQuery("SELECT COUNT(*) FROM Product").getSingleResult();
+        long singleResult = (long) session.createQuery("SELECT COUNT(*) FROM Product").getSingleResult();
         session.close();
         return singleResult;
     }
